version 0.4.5
This revision includes an extra argument to FactorModelCovariance
addressing GH issue #102, you can now specify also the factor covariance
together with the factor exposure and the idyosyncratic variances. Refer
to the documentation for details. Also, now most objects have a __repr__
method that prints out informations about them to the interpreter. That
will be used to store backtest results in future releases (which will
allow for smarter ways to optimize over hyperparameters repeatedly).