version 0.4.6
This release only contains a minor fix (see GH issue #104) with
the install requirements in setup.py. We have removed a limit on scipy
versioning that was meant to prevent conflict with cvxpy (a scipy release
in summer 2023 triggered a bug in cvxpy, which was promptly fixed,
but still presented an issue to users who didn't update).