version 0.4.7
This release (mostly) contains fixes to github issues #106 and #107.
The first solves a bug that happened when the user provides data
(e.g., r_hat for ReturnsForecast) to be used in a backtest with varying
universe, such as when there are IPOs in the backtest period. There was
no logic to only select the relevant part (columns, in that case) of the
user-provided data, and now there is, including multiple error checks.
The second solves a bug in the accounting of short positions (the book
is right, the code deviated from it). We changed CashReturn and
(Stocks)HoldingCost, which have now a much clearer (and better
documented) interface.