version 0.4.9
This revision fixes some bugs that were introduced while addressing
issue #106 (user provided data now gets sliced during a backtest
according to the trading universe at each point in time, since that
may change). Data-related errors are now more informative, and tell
the user whether data is missing at a certain point in time, some assets
are missing, or there are NaNs. We are mostly working on the
documentation now and are fixing inconsistencies between the book and
the code.