version 0.4.10
This revision contains various bug fixes, new better __repr__ (which is called when the user prints) and plot methods of BacktestResult, now both shown in the documentation in the Hello World Example page. Also, we include a first implementation of MarketSimulator.optimize_hyperparameters, which takes a policy defined with symbolic hyperparameters and optimizes their values for a user-specified back-test objective (by default Sharpe ratio). This is done very naïvely by simple greedy grid search of the hyper-parameter values.