version 1.0.1
This revision contains bug fixes and no new features.
An inefficient operation in TransactionCost.simulate (that
was introduced to comply with the new MarketData.serve
protocol) has been reworked (if you noticed a slowdown
in back-tests this probably fixes it). In SymbolData.update
the data quality checks that have been introduced in 1.0
have been adjusted to give less false positive error
logs.