version 1.0.3
This revision contains minor bug fixes and improvements
of the code quality (the total pylint score is now 9.7).

We now support Python 3.12.

Some aliases have been added to reflect names that were
defined in the original paper and had been changed recently.
We will in the coming weeks restore the original examples.

The documentation is growing and we have now exposed the core
internal classes through it.

Better error checks have been added, addressing GH issue #115
(when the provided returns are shorter than min_history, and
there can be no back-test, a clear error is now thrown).
The min_history parameter is now also an argument of
MarketSimulator (was only of the MarketData servers before).