version 1.1.0
This minor version includes numerous bug fixes and new features.
(We follow the semantic versioning specification, https://semver.org/ ,
so new features are only introduced at minor and major releases.)
The BackTestResult object now has a "logs" property which returns
all logs produced by Cvxportfolio objects, and other Python code that
might be called by user-defined objects, during a back-test. If running
in parallel, only logs from the process that runs the one back-test are
included. We also added the "online_usage" option to the market data
servers, and added to the examples a few strategies which we run online
every day (at US market open) committing their results to the code repository,
to certify them (those are not included in the code distribution, which
is very small in size). Then, we added a new "finalize_estimator" method
with the accompanying recursive tree method to all estimators, which we
plan to use to improve the caching logic (giving more options, such as
storage format, and opening it to third-party usage). Notable bug fixes
are the improved "__repr__" method of estimators, in particular
optimization-based policies (which now are well readable), a few (not
all) of the original paper examples have been translated to the stable API,
and we have improved documentation, docstrings, and pylint scores.